add custom hook for context
