+++ src/contexts/ContentViewContext.js
@@ -1,5 +1,33 @@
-import { createContext } from 'react';
+import { createContext, useState } from 'react';
+
+import Components from 'components/Constants/Components';
 
 const ContentViewContext = createContext();
 
-export default ContentViewContext;
+const ContentViewContextProvider = ({ children }) => {
+  const [displayedComponent, setDisplayedComponent] = useState(
+    Components.WELCOME
+  );
+  const [displayMap, setDisplayMap] = useState(false);
+  const [navbarColor, setNavbarColor] = useState('white');
+  const [navbarLeftButton, setNavbarLeftButton] = useState('');
+
+  const contentViewFunctions = {
+    displayedComponent,
+    setDisplayedComponent,
+    displayMap,
+    setDisplayMap,
+    navbarColor,
+    setNavbarColor,
+    navbarLeftButton,
+    setNavbarLeftButton,
+  };
+
+  return (
+    <ContentViewContext.Provider value={contentViewFunctions}>
+      {children}
+    </ContentViewContext.Provider>
+  );
+};
+
+export {ContentViewContext, ContentViewContextProvider};
